+++ src/product-service/handlers/getProductsById.ts
@@ -1,8 +1,8 @@
 // import { APIGatewayProxyHandler } from 'aws-lambda';
 import 'source-map-support/register';
 import { Client } from 'pg';
-import { dbOptions } from '../utils/dbOptions';
-import { corsHeaders } from '../utils/corsHeaders';
+import { dbOptions } from '../../utils/dbOptions';
+import { corsHeaders } from '../../utils/corsHeaders';
 
 export const getProductsById = async event => {
   console.log('Get product by id: ', event.pathParameters);
