+++ src/product-service/handlers/updateProduct.ts
@@ -1,8 +1,8 @@
 // import { APIGatewayProxyHandler } from 'aws-lambda';
 import 'source-map-support/register';
 import { Client } from 'pg';
-import { dbOptions } from '../utils/dbOptions';
-import { corsHeaders } from '../utils/corsHeaders';
+import { dbOptions } from '../../utils/dbOptions';
+import { corsHeaders } from '../../utils/corsHeaders';
 
 export const updateProduct = async event => {
   console.log('Update product: ', event.body);
