+++ src/product-service/tests/deleteProduct.test.ts
@@ -1,5 +1,5 @@
 import { deleteProduct } from '../handlers/deleteProduct';
-import { corsHeaders } from '../utils/corsHeaders';
+import { corsHeaders } from '../../utils/corsHeaders';
 
 test('should delete product', async () => {
   const product = await deleteProduct({ pathParameters: { productId: '6341365f-3f0d-4c07-aa0b-bf96f95ab426' }});
